+++ bika/lims/browser/analysisrequest/manage_analyses.py
@@ -111,7 +111,7 @@ class AnalysisRequestAnalysesView(BikaListingView):
         self.review_states = [
             {'id': 'default',
              'title': _('All'),
-             'contentFilter': {},
+             'contentFilter': {"inactive_state": "active"},
              'columns': columns,
              'transitions': [{'id': 'empty'}, ],  # none
              'custom_transitions': [{'id': 'save_analyses_button',
@@ -287,13 +287,6 @@ class AnalysisRequestAnalysesView(BikaListingView):
                 title='%s'>" % (
                     self.portal_url,
                     t(_("Accredited"))
-                )
-            if obj.getReportDryMatter():
-                after_icons += "<img\
-                src='%s/++resource++bika.lims.images/dry.png'\
-                title='%s'>" % (
-                    self.portal_url,
-                    t(_("Can be reported as dry matter"))
                 )
             if obj.getAttachmentOption() == 'r':
                 after_icons += "<img\