+++ bika/lims/browser/widgets/referenceresultswidget.py
@@ -91,11 +91,6 @@ class ReferenceResultsView(BikaListingView):
                 src='%s/++resource++bika.lims.images/accredited.png'\
                 title='%s'>" % (self.context.absolute_url(),
                                 _("Accredited"))
-            if service.getReportDryMatter():
-                after_icons += "<img\
-                src='%s/++resource++bika.lims.images/dry.png'\
-                title='%s'>" % (self.context.absolute_url(),
-                                _("Can be reported as dry matter"))
             if service.getAttachmentOption() == 'r':
                 after_icons += "<img\
                 src='%s/++resource++bika.lims.images/attach_reqd.png'\