+++ CHANGES.rst
@@ -6,16 +6,30 @@ Changelog
 
 **Added**
 
+- #777 Delay option for searches in reference widget combos
+- #806 Include Client ID when setting up ARReport on the IDServer
 
 **Changed**
 
 
 **Removed**
 
+- #800 Remove Dry Matter from tests
+- #779 Remove Dry Matter functionality
+
 
 **Fixed**
 
 - #726 404 Error raised when clicking Print Samples Sheets from within a client
+- #802 Remove Dry Matter remainders
+- #781 Delete Permission on ARs for Lab Managers
+- #784 Fix workflow state filter not kept when expanding categories in AS listing
+- #786 Fix inactive services listed in AR "Manage Analyses" forms
+- #775 Analyses on Analysis Requests are hyperlinked to their Worksheets
+- #769 Traceback when submitting duplicate when Duplicate Variation is not set
+- #771 Slow Searches in Listing Views
+- #774 When retracting an Analysis Requests its analyses are also retracted
+- #772 Improved UID check in API
 
 **Security**
 
@@ -34,6 +48,7 @@ Changelog
 
 **Changed**
 
+- #765 Department Filtering Improvements
 - #746 StringField to UIDReferenceField for Default Department of Lab Contact
 - #744 Updated WeasyPrint to 0.42.2
 - #694 Out of range/shoulders logic redux, ported to `api.analysis`